requirements.txt
requests
sqlalchemy
openai
google-generativeai
schedule
beautifulsoup4
python-dotenv
tenacity
httpx
PyYAML
feedparser
listparser
